Delete `views`, `likes` etc when fork. #40. Can give likes only to personal notes. #42 Don't open Bookmarks from navigation. #43 Bad encoding PDF (russian). #41 Add: migrations.

+++ noted/content/signals.py
@@ -7,5 +7,5 @@ from content.models import Note
 @receiver(post_delete, sender=Note)
 def free_source(sender, instance, **kwargs):
     """Delete :model:`Source` instance from the database if it has 0 notes."""
-    if not instance.source.notes.all():
+    if instance.source and not instance.source.notes.all():
         instance.source.delete()
